backend/requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
sentence-transformers
scikit-learn
pymysql
requests
python-dotenv
